+++ tests/test_contentsmanager.py
@@ -391,3 +391,64 @@ def test_preferred_format_allows_to_read_implicit_light_format(nb_file, tmpdir):
 
     # check that format (missing) is recognized as light
     assert 'py:light' in model['content']['metadata']['jupytext']['formats']
+
+
+@pytest.mark.parametrize('nb_file', list_notebooks('ipynb'))
+def test_save_in_auto_extension(nb_file, tmpdir):
+    # load notebook
+    nb = jupytext.readf(nb_file)
+    if 'language_info' not in nb.metadata:
+        return
+
+    auto_ext = nb.metadata['language_info']['file_extension']
+    if auto_ext == '.r':
+        auto_ext = '.R'
+    tmp_ipynb = 'notebook.ipynb'
+    tmp_script = 'notebook' + auto_ext
+
+    # create contents manager with default load format as percent
+    cm = jupytext.TextFileContentsManager()
+    cm.default_jupytext_formats = 'ipynb,auto'
+    cm.preferred_jupytext_formats_save = 'auto:percent'
+    cm.root_dir = str(tmpdir)
+
+    # save notebook
+    with mock.patch('jupytext.header.INSERT_AND_CHECK_VERSION_NUMBER', True):
+        cm.save(model=dict(type='notebook', content=nb), path=tmp_ipynb)
+
+    # check that text representation exists, and is in percent format
+    with open(str(tmpdir.join(tmp_script))) as stream:
+        assert read_format_from_metadata(stream.read(), auto_ext) == 'percent'
+
+
+@pytest.mark.parametrize('nb_file', list_notebooks('ipynb'))
+def test_save_in_pct_and_lgt_auto_extensions(nb_file, tmpdir):
+    # load notebook
+    nb = jupytext.readf(nb_file)
+    if 'language_info' not in nb.metadata:
+        return
+
+    auto_ext = nb.metadata['language_info']['file_extension']
+    if auto_ext == '.r':
+        auto_ext = '.R'
+    tmp_ipynb = 'notebook.ipynb'
+    tmp_pct_script = 'notebook.pct' + auto_ext
+    tmp_lgt_script = 'notebook.lgt' + auto_ext
+
+    # create contents manager with default load format as percent
+    cm = jupytext.TextFileContentsManager()
+    cm.default_jupytext_formats = 'ipynb,pct.auto,lgt.auto'
+    cm.preferred_jupytext_formats_save = 'pct.auto:percent,lgt.auto:light'
+    cm.root_dir = str(tmpdir)
+
+    # save notebook
+    with mock.patch('jupytext.header.INSERT_AND_CHECK_VERSION_NUMBER', True):
+        cm.save(model=dict(type='notebook', content=nb), path=tmp_ipynb)
+
+    # check that text representation exists in percent format
+    with open(str(tmpdir.join(tmp_pct_script))) as stream:
+        assert read_format_from_metadata(stream.read(), '.pct' + auto_ext) == 'percent'
+
+    # check that text representation exists in light format
+    with open(str(tmpdir.join(tmp_lgt_script))) as stream:
+        assert read_format_from_metadata(stream.read(), '.lgt' + auto_ext) == 'light'