+++ HISTORY.rst
@@ -12,6 +12,7 @@ Release History
 - Notebooks extensions can be prefixed with any prefix of at most three chars (#87)
 - Export of the same notebook to multiple formats is now supported. To export to all python formats, plus ``.ipynb`` and ``.md``, use ``"jupytext": {"formats": "ipynb,pct.py:percent,lgt.py:light,spx.py:sphinx,md"},``.
 - README includes a short section on how to extend ``light`` and ``percent`` formats to more languages (#61)
+- Jupytext's contents manager accepts the ``auto`` extension in ``default_jupytext_formats`` (#93)
 
 **BugFixes**